Code cleanup.

+++ Gravitas.Core/DeviceManager/Card/ICardManager.cs
@@ -1,5 +1,4 @@
 using Gravitas.Model.DomainModel.EmployeeRoles.DTO;
-using Gravitas.Model.Dto;
 using Node = Gravitas.Model.DomainModel.Node.TDO.Detail.Node;
 
 namespace Gravitas.Core.DeviceManager.Card
@@ -9,8 +8,8 @@ namespace Gravitas.Core.DeviceManager.Card
         CardReadResult GetTruckCardByOnGateReader(Node nodeDto);
         CardReadResult GetTruckCardByZebraReader(Node nodeDto);
         Model.DomainModel.Card.DAO.Card GetLaboratoryTrayOnTableReader(Node nodeDto);
-        bool IsMasterEmployeeCard(Model.DomainModel.Card.DAO.Card card, long nodeId);
-        bool IsLaboratoryEmployeeCard(Model.DomainModel.Card.DAO.Card card, long nodeId);
+        bool IsMasterEmployeeCard(Model.DomainModel.Card.DAO.Card card, int nodeId);
+        bool IsLaboratoryEmployeeCard(Model.DomainModel.Card.DAO.Card card, int nodeId);
         bool IsAdminCard(Model.DomainModel.Card.DAO.Card card, RolesDto employeeRoles = null);
         void SetRfidValidationDO(bool isValid, Node nodeDto);
         (CardReadResult card, bool input) GetTruckCardByZebraReaderDirection(Node nodeDto);
